Add stalemate detection, move check/stale mate detection to server
Moving check and stalemate detection to server greatly simplifies client code.
It also prevents a nefarious client from trying to win the game early.
Detection uses a server copy of the board to verify via library whether board is
in a state of checkmate or stalemate for the current player, and sends
a message to both clients on game end.