Add support for move history in client
This addresses issue #11, namely that move history is not displayed on the client. This
commit adds the latest chess library which adds support for SAN moves and displays the moves
in a table. This table is part of game info and is updated by a synchronize method every
time a move is applied by the player or downloaded in via sync from socket.io. The
styling needs some work but function-wise this is complete.